+++ src/hooks/useForm.js
@@ -1,6 +1,6 @@
-import React, { useState, useReducer } from "react";
+import { useState, useReducer, useMemo, useCallback, useEffect } from "react";
 
-const useForm = ({ forms }) => {
+const useForm = (mode) => {
   // stores value for all form inputs
   const [_formData, _setFormData] = useReducer(
     (data, { key, value, valid }) => {
@@ -10,8 +10,44 @@ const useForm = ({ forms }) => {
     },
     {}
   );
+
+  // data validity state
   const [valid, setValid] = useState(false);
-  return <form>{forms.map(() => {})}</form>;
+
+  // memoized form data object
+  const formData = useMemo(() => {
+    return { ..._formData };
+  }, [_formData]);
+
+  useEffect(() => {
+    let _valid = true;
+    Object.values(formData).forEach((data) => {
+      _valid = _valid && data.valid;
+    });
+    setValid(_valid);
+    console.log(`[In InputForm] ${JSON.stringify(formData)}, ${_valid}`);
+  }, [formData]);
+
+  // memoized setFormData
+  const setFormData = useCallback(
+    (key, value, valid) => {
+      _setFormData({ key, value, valid });
+      console.log(`[in useCallback] key => ${key}, value=> ${value}`);
+    },
+    [_setFormData]
+  );
+
+  // submit form handler
+  const submitForm = (e) => {
+    e.preventDefault();
+    console.log(`[in submit form] formdata => ${JSON.stringify(formData)}`);
+  };
+
+  return {
+    valid,
+    setFormData,
+    submitForm,
+  };
 };
 
 export default useForm;
